VORTEX-5590: Repackaging functions to separate into smaller packages.
