+++ open-sphere-base/analysis/src/main/java/io/opensphere/analysis/table/functions/statusbar/Max.java
@@ -1,33 +1,34 @@
-package io.opensphere.analysis.table.functions;
+package io.opensphere.analysis.table.functions.statusbar;
 
 import javax.swing.JTable;
 
 import io.opensphere.core.Toolbox;
 import io.opensphere.core.util.lang.NumberUtilities;
 
-/** Representation of Table Cell Summation. */
-public class Sum extends StatusBarFunction
+/** Representation of Table Selection Maximum. */
+public class Max extends StatusBarFunction
 {
     /**
-     * Constructs a Sum function.
+     * Constructs a Max function.
      *
      * @param toolbox The toolbox through which application state is accessed.
      */
-    public Sum(Toolbox toolbox)
+    public Max(Toolbox toolbox)
     {
-        super(toolbox, "%-10.3f", "Sum");
+        super(toolbox, "%-10.3f", "Maximum Value");
     }
 
     /**
-     * Sums the selected cells in a given table.
+     * Determines the maximum value of the selected cells in a given table.
      *
      * @override
+     * @return the maximum value, of 0.0 if nothing is selected
      */
     @SuppressWarnings("boxing")
     @Override
     public Number execute(JTable table)
     {
-        double sum = 0.0;
+        double max = 0.0;
 
         int[] rows = table.getSelectedRows();
         int[] cols = table.getSelectedColumns();
@@ -35,10 +36,10 @@ public class Sum extends StatusBarFunction
         {
             for (int c : cols)
             {
-                sum += NumberUtilities.parseDouble(table.getValueAt(r, c), 0.0);
+                max = Math.max(max, NumberUtilities.parseDouble(table.getValueAt(r, c), 0.0));
             }
         }
 
-        return sum;
+        return max;
     }
 }
