+++ open-sphere-base/analysis/src/main/java/io/opensphere/analysis/table/functions/statusbar/Min.java
@@ -1,4 +1,4 @@
-package io.opensphere.analysis.table.functions;
+package io.opensphere.analysis.table.functions.statusbar;
 
 import java.util.ArrayList;
 import java.util.List;
